Refactor client config API

+++ sdk-client/src/main/java/com/digitalpetri/opcua/sdk/client/fsm/states/CreateAndActivate.java
@@ -181,7 +181,7 @@ public class CreateAndActivate implements SessionState {
                 client.getConfig().getSessionName().get(),
                 clientNonce,
                 clientCertificate,
-                client.getConfig().getSessionTimeout(),
+                client.getConfig().getSessionTimeout().doubleValue(),
                 client.getConfig().getMaxResponseMessageSize());
 
         logger.debug("Sending CreateSessionRequest...");
